Move code responsible for formatting responses to different file
