+++ src/index.ts
@@ -1,8 +1,8 @@
-import deepReplace from './helpers/deepReplace'
+import { stringifiedRequestAttributes } from './formatting'
 
 export default class ApolloLogExtension {
     requestDidStart(request: any): void {
-        console.log(prefix() + stringifiedRequestAttributes(request))
+        console.log(stringifiedRequestAttributes(request))
     }
   
     willSendResponse(object: any): void {
@@ -12,32 +12,4 @@ export default class ApolloLogExtension {
     parsingDidStart(r: any): void {}
     validationDidStart(): void {}
     executionDidStart(r: any): void {}
-}
-
-
-const prefix = (): string => {
-    return `[${Date.now()}] `
-}
-
-const filterPasswordFromVariables = (variables: { password?: string }) => {
-    deepReplace(variables, "password", "[FILTERED]")
-
-    return variables
-}
-
-const stringifiedRequestAttributes = ({ headers, variables, queryString, operationName }: request): string => {
-    const filteredVariables = filterPasswordFromVariables(variables)
-
-    const stringifiedHeaders = JSON.stringify(headers)
-    const stringifiedVariables = JSON.stringify(filteredVariables)
-    const stringifiedQueryString = JSON.stringify(queryString).replace(/\s/g, '').replace(/\\n/g, " ")
-
-    return `${operationName}\nHeaders: ${stringifiedHeaders}\nQueryString: ${stringifiedQueryString}\nVariables: ${stringifiedVariables}`
-}
-
-export interface request {
-    headers: object
-    queryString: any
-    operationName: string
-    variables: object
 }
